+++ zrml/parimutuel/src/tests/claim.rs
@@ -27,181 +27,6 @@ use zeitgeist_primitives::types::{Asset, MarketStatus, MarketType, OutcomeReport
 use zrml_market_commons::{Error as MError, Markets};
 
 #[test]
-fn buy_emits_event() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_ok!(Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount));
-
-        let amount_minus_fees = 9900000000;
-        let fees = 100000000;
-        assert_eq!(amount, amount_minus_fees + fees);
-
-        System::assert_last_event(
-            Event::OutcomeBought { market_id, buyer: ALICE, asset, amount_minus_fees, fees }.into(),
-        );
-    });
-}
-
-#[test]
-fn buy_balances_change_correctly() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.status = MarketStatus::Active;
-        market.creator = MARKET_CREATOR;
-        Markets::<Runtime>::insert(market_id, market.clone());
-
-        let base_asset = market.base_asset;
-
-        let free_alice_before = AssetManager::free_balance(base_asset, &ALICE);
-        let free_creator_before = AssetManager::free_balance(base_asset, &market.creator);
-        let free_pot_before =
-            AssetManager::free_balance(base_asset, &Parimutuel::pot_account(market_id));
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_ok!(Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount));
-
-        let amount_minus_fees = 9900000000;
-        let fees = 100000000;
-        assert_eq!(amount, amount_minus_fees + fees);
-
-        assert_eq!(AssetManager::free_balance(base_asset, &ALICE), free_alice_before - amount);
-        assert_eq!(
-            AssetManager::free_balance(base_asset, &Parimutuel::pot_account(market_id))
-                - free_pot_before,
-            amount_minus_fees
-        );
-        assert_eq!(AssetManager::free_balance(asset, &ALICE), amount_minus_fees);
-        assert_eq!(
-            AssetManager::free_balance(base_asset, &market.creator) - free_creator_before,
-            fees
-        );
-    });
-}
-
-#[test]
-fn buy_fails_if_asset_not_parimutuel_share() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market.clone());
-
-        let asset = Asset::CategoricalOutcome(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_noop!(
-            Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount),
-            Error::<Runtime>::NotParimutuelOutcome
-        );
-    });
-}
-
-#[test_case(ScoringRule::CPMM; "cpmm")]
-#[test_case(ScoringRule::Orderbook; "orderbook")]
-#[test_case(ScoringRule::Lmsr; "lmsr")]
-#[test_case(ScoringRule::RikiddoSigmoidFeeMarketEma; "rikiddo sigmoid fee market ema")]
-fn buy_fails_if_invalid_scoring_rule(scoring_rule: ScoringRule) {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.status = MarketStatus::Active;
-        market.scoring_rule = scoring_rule;
-
-        Markets::<Runtime>::insert(market_id, market.clone());
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_noop!(
-            Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount),
-            Error::<Runtime>::InvalidScoringRule
-        );
-    });
-}
-
-#[test_case(MarketStatus::Proposed; "proposed")]
-#[test_case(MarketStatus::Suspended; "suspended")]
-#[test_case(MarketStatus::Closed; "closed")]
-#[test_case(MarketStatus::CollectingSubsidy; "collecting subsidy")]
-#[test_case(MarketStatus::InsufficientSubsidy; "insufficient subsidy")]
-#[test_case(MarketStatus::Reported; "reported")]
-#[test_case(MarketStatus::Disputed; "disputed")]
-fn buy_fails_if_market_status_is_not_active(status: MarketStatus) {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.status = status;
-        market.scoring_rule = ScoringRule::Parimutuel;
-
-        Markets::<Runtime>::insert(market_id, market.clone());
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_noop!(
-            Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount),
-            Error::<Runtime>::MarketIsNotActive
-        );
-    });
-}
-
-#[test]
-fn buy_fails_if_insufficient_balance() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market.clone());
-
-        let free_alice = AssetManager::free_balance(market.base_asset, &ALICE);
-        AssetManager::slash(market.base_asset, &ALICE, free_alice);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_noop!(
-            Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount),
-            Error::<Runtime>::InsufficientBalance
-        );
-    });
-}
-
-#[test]
-fn buy_fails_if_below_minimum_bet_size() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market.clone());
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get() - 1;
-        assert_noop!(
-            Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount),
-            Error::<Runtime>::AmountTooSmall
-        );
-    });
-}
-
-#[test]
-fn buy_fails_if_market_does_not_exist() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_noop!(
-            Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount),
-            MError::<Runtime>::MarketDoesNotExist
-        );
-    });
-}
-
-#[test]
 fn claim_rewards_emits_event() {
     ExtBuilder::default().build().execute_with(|| {
         let market_id = 0;
@@ -329,26 +154,6 @@ fn claim_rewards_categorical_changes_balances_correctly() {
         assert_eq!(
             AssetManager::free_balance(market.base_asset, &Parimutuel::pot_account(market_id)),
             0
-        );
-    });
-}
-
-#[test]
-fn buy_fails_if_market_type_is_scalar() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        let range: RangeInclusive<u128> = 0..=100;
-        market.market_type = MarketType::Scalar(range);
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount =
-            <Runtime as Config>::MinBetSize::get() + <Runtime as Config>::MinBetSize::get();
-        assert_noop!(
-            Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount),
-            Error::<Runtime>::NotCategorical
         );
     });
 }
@@ -565,199 +370,6 @@ fn claim_refunds_works() {
         assert_eq!(
             AssetManager::free_balance(market.base_asset, &Parimutuel::pot_account(market_id)),
             0
-        );
-    });
-}
-
-#[test]
-fn refund_fails_if_not_parimutuel_outcome() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.market_type = MarketType::Categorical(10u16);
-        market.resolved_outcome = Some(OutcomeReport::Categorical(0u16));
-        market.status = MarketStatus::Resolved;
-        Markets::<Runtime>::insert(market_id, market);
-
-        assert_noop!(
-            Parimutuel::claim_refunds(
-                RuntimeOrigin::signed(ALICE),
-                Asset::CategoricalOutcome(market_id, 0u16)
-            ),
-            Error::<Runtime>::NotParimutuelOutcome
-        );
-    });
-}
-
-#[test_case(MarketStatus::Active; "active")]
-#[test_case(MarketStatus::Proposed; "proposed")]
-#[test_case(MarketStatus::Suspended; "suspended")]
-#[test_case(MarketStatus::Closed; "closed")]
-#[test_case(MarketStatus::CollectingSubsidy; "collecting subsidy")]
-#[test_case(MarketStatus::InsufficientSubsidy; "insufficient subsidy")]
-#[test_case(MarketStatus::Reported; "reported")]
-#[test_case(MarketStatus::Disputed; "disputed")]
-fn refund_fails_if_market_not_resolved(status: MarketStatus) {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.market_type = MarketType::Categorical(10u16);
-        market.status = status;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        assert_noop!(
-            Parimutuel::claim_refunds(RuntimeOrigin::signed(ALICE), asset),
-            Error::<Runtime>::MarketIsNotResolvedYet
-        );
-    });
-}
-
-#[test_case(ScoringRule::CPMM; "cpmm")]
-#[test_case(ScoringRule::Orderbook; "orderbook")]
-#[test_case(ScoringRule::Lmsr; "lmsr")]
-#[test_case(ScoringRule::RikiddoSigmoidFeeMarketEma; "rikiddo sigmoid fee market ema")]
-fn refund_fails_if_invalid_scoring_rule(scoring_rule: ScoringRule) {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.market_type = MarketType::Categorical(10u16);
-        market.resolved_outcome = Some(OutcomeReport::Categorical(0u16));
-        market.status = MarketStatus::Resolved;
-        // invalid scoring rule
-        market.scoring_rule = scoring_rule;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        assert_noop!(
-            Parimutuel::claim_refunds(RuntimeOrigin::signed(ALICE), asset),
-            Error::<Runtime>::InvalidScoringRule
-        );
-    });
-}
-
-#[test]
-fn refund_fails_if_invalid_outcome_asset() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.market_type = MarketType::Categorical(10u16);
-        market.resolved_outcome = Some(OutcomeReport::Categorical(0u16));
-        market.status = MarketStatus::Resolved;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 20u16);
-        assert_noop!(
-            Parimutuel::claim_refunds(RuntimeOrigin::signed(ALICE), asset),
-            Error::<Runtime>::InvalidOutcomeAsset
-        );
-    });
-}
-
-#[test]
-fn refund_fails_if_no_resolved_outcome() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.market_type = MarketType::Categorical(10u16);
-        market.status = MarketStatus::Resolved;
-        market.resolved_outcome = None;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        assert_noop!(
-            Parimutuel::claim_refunds(RuntimeOrigin::signed(ALICE), asset),
-            Error::<Runtime>::NoResolvedOutcome
-        );
-    });
-}
-
-#[test]
-fn refund_fails_if_refund_not_allowed() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.market_type = MarketType::Categorical(10u16);
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_ok!(Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount));
-
-        let mut market = Markets::<Runtime>::get(market_id).unwrap();
-        market.resolved_outcome = Some(OutcomeReport::Categorical(0u16));
-        market.status = MarketStatus::Resolved;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        assert_noop!(
-            Parimutuel::claim_refunds(RuntimeOrigin::signed(ALICE), asset),
-            Error::<Runtime>::RefundNotAllowed
-        );
-    });
-}
-
-#[test]
-fn refund_fails_if_refundable_balance_is_zero() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.market_type = MarketType::Categorical(10u16);
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_ok!(Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount));
-
-        let mut market = Markets::<Runtime>::get(market_id).unwrap();
-        market.resolved_outcome = Some(OutcomeReport::Categorical(1u16));
-        market.status = MarketStatus::Resolved;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        assert_ok!(Parimutuel::claim_refunds(RuntimeOrigin::signed(ALICE), asset));
-
-        // already refunded above
-        assert_noop!(
-            Parimutuel::claim_refunds(RuntimeOrigin::signed(ALICE), asset),
-            Error::<Runtime>::RefundableBalanceIsZero
-        );
-    });
-}
-
-#[test]
-fn refund_emits_event() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>();
-        market.market_type = MarketType::Categorical(10u16);
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        let amount = <Runtime as Config>::MinBetSize::get();
-        assert_ok!(Parimutuel::buy(RuntimeOrigin::signed(ALICE), asset, amount));
-
-        let mut market = Markets::<Runtime>::get(market_id).unwrap();
-        market.resolved_outcome = Some(OutcomeReport::Categorical(1u16));
-        market.status = MarketStatus::Resolved;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset = Asset::ParimutuelShare(market_id, 0u16);
-        assert_ok!(Parimutuel::claim_refunds(RuntimeOrigin::signed(ALICE), asset));
-
-        let amount_minus_fees = amount - (Percent::from_percent(1) * amount);
-
-        System::assert_last_event(
-            Event::BalanceRefunded {
-                market_id,
-                asset,
-                refunded_balance: amount_minus_fees,
-                sender: ALICE,
-            }
-            .into(),
         );
     });
 }
